@@ -24,23 +42,24 @@
     <name>swotlyzer</name>
     <description>swotlyzer</description>
     <properties>
-        <java.version>11</java.version>
-        <spring-boot.version>2.7.0</spring-boot.version>
+        <java.version>17</java.version>
+        <spring-boot.version>3.0.0</spring-boot.version>
         <spring-fox-swagger2.version>2.9.2</spring-fox-swagger2.version>
         <spring-fox-swagger-ui.version>2.9.2</spring-fox-swagger-ui.version>
-        <jsonwebtoken.version>0.9.1</jsonwebtoken.version>
+        <jsonwebtoken.version>0.11.5</jsonwebtoken.version>
         <jaxb-api.version>2.3.0</jaxb-api.version>
         <lombok.version>1.18.24</lombok.version>
         <spring-dotenv.version>2.5.4</spring-dotenv.version>
-        <spring-cloud.version>2021.0.2</spring-cloud.version>
-        <eureka.version>3.1.3</eureka.version>
-        <spring-cloud-gateway.version>3.1.3</spring-cloud-gateway.version>
-        <h2.version>1.4.200</h2.version>
+        <spring-cloud.version>2022.0.0-RC3</spring-cloud.version>
+        <eureka.version>4.0.0-RC3</eureka.version>
+        <cloud-config.version>4.0.0-RC3</cloud-config.version>
+        <spring-cloud-gateway.version>4.0.0-RC3</spring-cloud-gateway.version>
+        <h2.version>2.1.214</h2.version>
         <mock-server.version>4.0.1</mock-server.version>
         <ok-http.version>4.0.1</ok-http.version>
         <mockito-inline.version>4.7.0</mockito-inline.version>
-        <spring-kafka.version>2.9.0</spring-kafka.version>
-        <flyway.version>8.1.0</flyway.version>
+        <spring-kafka.version>3.0.0</spring-kafka.version>
+        <flyway.version>9.8.3</flyway.version>
         <spring-cloud-sleuth.version>3.1.5</spring-cloud-sleuth.version>
     </properties>
     <dependencyManagement>
