@@ -1,11 +1,9 @@
 package com.microservices.swotlyzer.mail.sender;
 
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.SpringApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@EnableEurekaClient
 public class MailSenderApplication {
 
     public static void main(String[] args) {
